Move azure.HandleCheckRunEvent into api.go
along with its test. One benefit is that api/azure/webhook_test.go now
no longer needs to be in a separate package to avoid cyclic dependency
and a bunch of regexps no longer need to be exported.
