+++ api/azure/webhook_test.go
@@ -4,115 +4,14 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-package azure_test
+package azure
 
 import (
-	"context"
 	"encoding/json"
-	"net/http"
-	"net/http/httptest"
-	"net/url"
-	"strings"
 	"testing"
 
-	"github.com/golang/mock/gomock"
-	"github.com/google/go-github/github"
-	logrustest "github.com/sirupsen/logrus/hooks/test"
 	"github.com/stretchr/testify/assert"
-
-	"github.com/web-platform-tests/wpt.fyi/api/azure"
-	"github.com/web-platform-tests/wpt.fyi/api/azure/mock_azure"
-	"github.com/web-platform-tests/wpt.fyi/shared"
-	"github.com/web-platform-tests/wpt.fyi/shared/sharedtest"
 )
-
-func TestHandleCheckRunEvent(t *testing.T) {
-	mockCtrl := gomock.NewController(t)
-	defer mockCtrl.Finish()
-
-	sha := strings.Repeat("0123456789", 4)
-	detailsURL := "https://dev.azure.com/web-platform-tests/b14026b4-9423-4454-858f-bf76cf6d1faa/_build/results?buildId=123"
-
-	id := azure.PipelinesAppID
-	chrome := "chrome"
-	completed := "completed"
-	created := "created"
-	repoName := "wpt"
-	repoOwner := "web-platform-tests"
-	sender := "lukebjerring"
-	event := &github.CheckRunEvent{
-		Action: &created,
-		CheckRun: &github.CheckRun{
-			App:     &github.App{ID: &id},
-			Name:    &chrome,
-			Status:  &completed,
-			HeadSHA: &sha,
-		},
-		Repo: &github.Repository{
-			Name:  &repoName,
-			Owner: &github.User{Login: &repoOwner},
-		},
-		Sender: &github.User{Login: &sender},
-	}
-
-	event.CheckRun.DetailsURL = &detailsURL
-
-	artifact := azure.BuildArtifact{Name: "results"}
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		switch r.URL.Path {
-		case "/upload":
-			username, password, ok := r.BasicAuth()
-			assert.True(t, ok)
-			assert.Equal(t, username, "azure")
-			assert.Equal(t, password, "123")
-			w.WriteHeader(200)
-		case "/123/artifacts":
-			artifacts := azure.BuildArtifacts{
-				Count: 1,
-				Value: []azure.BuildArtifact{
-					artifact,
-				},
-			}
-			bytes, _ := json.Marshal(artifacts)
-			w.Write(bytes)
-		default:
-			assert.FailNow(t, "Invalid spoofed request "+r.URL.String())
-		}
-	}))
-	defer server.Close()
-
-	build := azure.Build{
-		TriggerInfo: azure.BuildTriggerInfo{
-			SourceBranch: "master",
-		},
-	}
-
-	azureAPI := mock_azure.NewMockAPI(mockCtrl)
-	serverURL, _ := url.Parse(server.URL)
-	azureAPI.EXPECT().GetAzureArtifactsURL(repoOwner, repoName, int64(123)).Return(server.URL + "/123/artifacts")
-	azureAPI.EXPECT().GetBuild(repoOwner, repoName, int64(123)).Return(&build)
-
-	aeAPI := sharedtest.NewMockAppEngineAPI(mockCtrl)
-	aeAPI.EXPECT().GetVersionedHostname().AnyTimes().Return(serverURL.Host)
-	uploadURL, _ := url.Parse(server.URL + "/upload")
-	aeAPI.EXPECT().GetResultsUploadURL().Return(uploadURL)
-	aeAPI.EXPECT().GetUploader("azure").Return(shared.Uploader{Username: "azure", Password: "123"}, nil)
-	aeAPI.EXPECT().GetHTTPClient().AnyTimes().Return(server.Client())
-	aeAPI.EXPECT().GetSlowHTTPClient(gomock.Any()).AnyTimes().Return(server.Client(), func() {})
-
-	log, hook := logrustest.NewNullLogger()
-	ctx := context.WithValue(context.Background(), shared.DefaultLoggerCtxKey(), log)
-	aeAPI.EXPECT().Context().AnyTimes().Return(ctx)
-	processed, err := azure.HandleCheckRunEvent(azureAPI, aeAPI, event)
-	if err != nil {
-		assert.FailNow(t, "Error isn't nil", err.Error())
-	}
-	assert.True(t, processed)
-	if len(hook.Entries) < 1 {
-		assert.FailNow(t, "No logging was found")
-	}
-	assert.Contains(t, hook.Entries[0].Message, "/123/")
-}
 
 const artifactsJSON = `{
 	"count": 2,
@@ -144,7 +43,7 @@ const artifactsJSON = `{
 }`
 
 func TestParses(t *testing.T) {
-	var artifacts azure.BuildArtifacts
+	var artifacts BuildArtifacts
 	err := json.Unmarshal([]byte(artifactsJSON), &artifacts)
 	assert.Nil(t, err)
 	assert.Equal(t, int64(2), artifacts.Count)
@@ -156,31 +55,31 @@ func TestParses(t *testing.T) {
 
 func TestArtifactRegexes(t *testing.T) {
 	// Names before https://github.com/web-platform-tests/wpt/pull/15110
-	assert.True(t, azure.MasterRegex.MatchString("results"))
-	assert.True(t, azure.PRHeadRegex.MatchString("affected-tests"))
-	assert.True(t, azure.PRBaseRegex.MatchString("affected-tests-without-changes"))
+	assert.True(t, masterRegex.MatchString("results"))
+	assert.True(t, prHeadRegex.MatchString("affected-tests"))
+	assert.True(t, prBaseRegex.MatchString("affected-tests-without-changes"))
 
 	// Names after https://github.com/web-platform-tests/wpt/pull/15110
-	assert.True(t, azure.MasterRegex.MatchString("edge-results"))
-	assert.True(t, azure.PRHeadRegex.MatchString("safari-preview-affected-tests"))
-	assert.True(t, azure.PRBaseRegex.MatchString("safari-preview-affected-tests-without-changes"))
+	assert.True(t, masterRegex.MatchString("edge-results"))
+	assert.True(t, prHeadRegex.MatchString("safari-preview-affected-tests"))
+	assert.True(t, prBaseRegex.MatchString("safari-preview-affected-tests-without-changes"))
 
 	// Don't accept the other order
-	assert.False(t, azure.MasterRegex.MatchString("results-edge"))
+	assert.False(t, masterRegex.MatchString("results-edge"))
 
 	// Don't accept any string ending with the right pattern
-	assert.False(t, azure.MasterRegex.MatchString("nodashresults"))
+	assert.False(t, masterRegex.MatchString("nodashresults"))
 
 	// Base and Head could be confused with substring matching
-	assert.False(t, azure.PRBaseRegex.MatchString("affected-tests"))
-	assert.False(t, azure.PRHeadRegex.MatchString("affected-tests-without-changes"))
+	assert.False(t, prBaseRegex.MatchString("affected-tests"))
+	assert.False(t, prHeadRegex.MatchString("affected-tests-without-changes"))
 }
 
 func TestEpochBranchesRegex(t *testing.T) {
-	assert.True(t, azure.EpochBranchesRegex.MatchString("refs/heads/epochs/twelve_hourly"))
-	assert.True(t, azure.EpochBranchesRegex.MatchString("refs/heads/epochs/six_hourly"))
-	assert.True(t, azure.EpochBranchesRegex.MatchString("refs/heads/epochs/weekly"))
-	assert.True(t, azure.EpochBranchesRegex.MatchString("refs/heads/epochs/daily"))
+	assert.True(t, epochBranchesRegex.MatchString("refs/heads/epochs/twelve_hourly"))
+	assert.True(t, epochBranchesRegex.MatchString("refs/heads/epochs/six_hourly"))
+	assert.True(t, epochBranchesRegex.MatchString("refs/heads/epochs/weekly"))
+	assert.True(t, epochBranchesRegex.MatchString("refs/heads/epochs/daily"))
 
-	assert.False(t, azure.EpochBranchesRegex.MatchString("refs/heads/weekly"))
+	assert.False(t, epochBranchesRegex.MatchString("refs/heads/weekly"))
 }
